// web-application/src/types-openapi/api/models/UpdateSupplierRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateSupplierRequest
 */
export interface UpdateSupplierRequest {
    /**
     * 
     * @type {number}
     * @memberof UpdateSupplierRequest
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    phone: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    address?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    city?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    district?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    ward?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateSupplierRequest
     */
    notes?: string | null;
}

/**
 * Check if a given object implements the UpdateSupplierRequest interface.
 */
export function instanceOfUpdateSupplierRequest(value: object): value is UpdateSupplierRequest {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('phone' in value) || value['phone'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    return true;
}

export function UpdateSupplierRequestFromJSON(json: any): UpdateSupplierRequest {
    return UpdateSupplierRequestFromJSONTyped(json, false);
}

export function UpdateSupplierRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateSupplierRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'phone': json['phone'],
        'email': json['email'],
        'address': json['address'] == null ? undefined : json['address'],
        'city': json['city'] == null ? undefined : json['city'],
        'district': json['district'] == null ? undefined : json['district'],
        'ward': json['ward'] == null ? undefined : json['ward'],
        'notes': json['notes'] == null ? undefined : json['notes'],
    };
}

export function UpdateSupplierRequestToJSON(json: any): UpdateSupplierRequest {
    return UpdateSupplierRequestToJSONTyped(json, false);
}

export function UpdateSupplierRequestToJSONTyped(value?: UpdateSupplierRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'phone': value['phone'],
        'email': value['email'],
        'address': value['address'],
        'city': value['city'],
        'district': value['district'],
        'ward': value['ward'],
        'notes': value['notes'],
    };
}